requirements.txt
aiogram
pillow
moviepy
lottie[all]
